requirements.txt
aiohttp==3.5.4
amqp==2.4.0
async-timeout==3.0.1
attrs==18.2.0
billiard==3.5.0.5
celery==4.2.1
certifi==2018.11.29
chardet==3.0.4
Django==2.1.5
django-celery-beat==1.4.0
django-environ==0.4.5
django-model-utils==3.1.2
django-timezone-field==3.0
djangorestframework==3.9.1
future==0.17.1
idna==2.8
idna-ssl==1.1.0
kombu==4.2.2.post1
multidict==4.5.2
psycopg2-binary==2.7.6.1
python-crontab==2.3.6
python-dateutil==2.7.5
python-telegram-bot==3.2.0
pytz==2018.9
redis==3.0.1
requests==2.21.0
six==1.12.0
telepot==12.7
typing-extensions==3.7.2
urllib3==1.24.1
vine==1.2.0
yarl==1.3.0
